Remove SF Pro fonts from test target
